+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-- Made the activation range of coffins from Trailier Tales configurable and set it to be much lower by default.
-- Wooden and stone spears are now disabled when Affogato Early Game Changes is enabled.
-- Improvements to the random structure chest placer.
+Affogato Early Game Changes fixes:
+- Wooden tools now have unique behavior and item descriptions.
+- Bonus chest now has flint tools instead of stone/wooden tools.
+- Wooden and stone spears no longer show up in the creative menu.
+
+Other changes:
+- Disabled BBE chest optimizations for stone chests from Wilder Wild if both mods are installed.
+- Updated and fixed default loot entry replacement settings.
+- Updated some item descriptions.